+++ intuita-webview/src/codemodList/CodemodNodeRenderer/style.module.css
@@ -105,9 +105,3 @@
 	font-size: 16px;
 	cursor: pointer;
 }
-
-.action {
-	margin-left: 3px;
-	display: flex;
-	align-items: center;
-}
