+++ intuita-webview/src/codemodList/TreeView/index.tsx
@@ -1,254 +1,67 @@
-import { ReactNode, useCallback, useEffect, useReducer, useState } from 'react';
-import Tree from './Tree';
-import TreeItem from './TreeItem';
-import {
-	RunCodemodsCommand,
-	CodemodTreeNode,
-	CodemodHash,
-	WebviewMessage,
-} from '../../shared/types';
-import { ReactComponent as CaseIcon } from '../../assets/case.svg';
-import { ReactComponent as BlueLightBulbIcon } from '../../assets/bluelightbulb.svg';
-import { vscode } from '../../shared/utilities/vscode';
-import styles from './style.module.css';
-import cn from 'classnames';
-import { useProgressBar } from '../useProgressBar';
+import { useCallback, useState } from 'react';
 import { VSCodeButton } from '@vscode/webview-ui-toolkit/react';
-import { SEARCH_QUERY_MIN_LENGTH } from '../../shared/SearchBar';
+
+import { RunCodemodsCommand, CodemodHash } from '../../shared/types';
 import Popover from '../../shared/Popover';
+import { vscode } from '../../shared/utilities/vscode';
+
+import { useProgressBar } from '../useProgressBar';
+
+import {
+	CodemodNode,
+	CodemodNodeHashDigest,
+	CodemodTree,
+} from '../../../../src/selectors/selectCodemodTree';
+
+import { IntuitaTreeView } from '../../intuitaTreeView';
+import { getCodemodNodeRenderer } from '../CodemodNodeRenderer';
+
+import s from './style.module.css';
 
 type Props = Readonly<{
-	node: CodemodTreeNode;
+	tree: CodemodTree;
 	autocompleteItems: string[];
-	openedIds: ReadonlySet<CodemodHash>;
-	nodeIds: ReadonlyArray<CodemodHash>;
-	nodesByDepth: ReadonlyArray<ReadonlyArray<CodemodTreeNode>>;
-	focusedId: CodemodHash | null;
-	searchQuery: string;
-	screenWidth: number | null;
+	rootPath: string;
 }>;
 
-export const containsSearchedCodemod = (
-	node: CodemodTreeNode,
-	searchQuery: string,
-	set: Set<CodemodHash>,
-): boolean => {
-	const lowerCaseSearchQuery = searchQuery.toLowerCase().trim();
-	if (
-		node.kind === 'codemodItem' &&
-		(node.uri.toLowerCase().includes(lowerCaseSearchQuery) ||
-			node.label.toLowerCase().includes(lowerCaseSearchQuery))
-	) {
-		set.add(node.id);
-		return true;
-	}
-	let someChildContains = false;
-	node.children.forEach((childNode) => {
-		const result = containsSearchedCodemod(childNode, searchQuery, set);
-		if (result) {
-			someChildContains = true;
-		}
-	});
-
-	if (someChildContains) {
-		set.add(node.id);
-	}
-
-	return someChildContains;
-};
-
-export const containsCodemodHashDigest = (
-	node: CodemodTreeNode,
-	codemodHashDigest: CodemodHash,
-	set: Set<CodemodHash>,
-): boolean => {
-	if (node.id === codemodHashDigest) {
-		return true;
-	}
-
-	const someChildContains = node.children.some((childNode) =>
-		containsCodemodHashDigest(childNode, codemodHashDigest, set),
-	);
-
-	if (someChildContains) {
-		set.add(node.id);
-	}
-
-	return someChildContains;
-};
-
-const getIcon = (iconName: string | null, open: boolean): ReactNode => {
-	if (iconName === 'case.svg') {
-		return <CaseIcon />;
-	}
-
-	if (iconName === 'folder.svg') {
-		return (
-			<span
-				className={cn(
-					'codicon',
-					!open ? 'codicon-folder' : 'codicon-folder-opened',
-				)}
-			/>
-		);
-	}
-
-	return <BlueLightBulbIcon />;
-};
-
-const handleDoubleClick = (node: CodemodTreeNode) => {
-	if (!node.doubleClickCommand) {
-		return;
-	}
-
+const handleDoubleClick = (node: CodemodNode) => {
 	vscode.postMessage({
 		kind: 'webview.command',
-		value: node.doubleClickCommand,
-	});
-};
-
-const handleClick = (node: CodemodTreeNode) => {
-	if (!node.command) {
-		return;
-	}
-
-	vscode.postMessage({
-		kind: 'webview.command',
-		value: node.command,
-	});
-};
-
-type State = Readonly<{
-	node: CodemodTreeNode;
-	openedIds: ReadonlySet<CodemodHash>;
-	focusedId: CodemodHash | null;
-}>;
-
-type InitializerArgument = State;
-
-type Action = Readonly<{
-	kind: 'focus' | 'flip' | 'open' | 'close';
-	id: CodemodHash;
-}>;
-
-const reducer = (state: State, action: Action): State => {
-	const openedIds = new Set(state.openedIds);
-
-	if (action.kind === 'focus') {
-		containsCodemodHashDigest(state.node, action.id, openedIds);
-
-		return {
-			node: state.node,
-			openedIds,
-			focusedId: action.id,
-		};
-	}
-
-	if (action.kind === 'flip') {
-		if (openedIds.has(action.id)) {
-			openedIds.delete(action.id);
-		} else {
-			openedIds.add(action.id);
-		}
-
-		return {
-			node: state.node,
-			openedIds,
-			focusedId: action.id,
-		};
-	}
-
-	if (action.kind === 'open') {
-		openedIds.add(action.id);
-
-		return {
-			node: state.node,
-			openedIds,
-			focusedId: action.id,
-		};
-	}
-
-	if (action.kind === 'close') {
-		if (openedIds.has(action.id)) {
-			openedIds.delete(action.id);
-		}
-
-		return {
-			node: state.node,
-			openedIds,
-			focusedId: action.id,
-		};
-	}
-
-	return state;
-};
-
-const initializer = ({
-	node,
-	focusedId,
-	openedIds,
-}: InitializerArgument): State => {
-	const newOpenedIds = new Set([...openedIds, node.id]);
-
-	if (focusedId !== null) {
-		containsCodemodHashDigest(node, focusedId, newOpenedIds);
-	}
-
-	return {
-		node,
-		openedIds: newOpenedIds,
-		focusedId,
-	};
-};
-
-const TreeView = ({
-	node,
-	autocompleteItems,
-	openedIds,
-	focusedId,
-	searchQuery,
-	nodeIds,
-	nodesByDepth,
-}: Props) => {
-	const rootPath = node.label;
-	const userSearchingCodemod = searchQuery.length >= SEARCH_QUERY_MIN_LENGTH;
-	const [hashesForSearch, setHashesForSearch] = useState<Set<CodemodHash>>(
-		new Set(),
-	);
-	const [state, dispatch] = useReducer(
-		reducer,
-		{
-			node,
-			focusedId,
-			openedIds,
+		value: {
+			title: 'Show codemod metadata',
+			command: 'intuita.showCodemodMetadata',
+			arguments: [node.hashDigest],
 		},
-		initializer,
-	);
+	});
+};
 
-	useEffect(() => {
-		if (searchQuery.length < SEARCH_QUERY_MIN_LENGTH) {
-			setHashesForSearch(new Set());
-			return;
-		}
-		const set = new Set<CodemodHash>();
-		containsSearchedCodemod(node, searchQuery, set);
+const onFocus = (hashDigest: CodemodNodeHashDigest) => {
+	vscode.postMessage({
+		kind: 'webview.global.selectCodemodNodeHashDigest',
+		selectedCodemodNodeHashDigest: hashDigest,
+	});
+};
 
-		setHashesForSearch(set);
-	}, [node, searchQuery]);
+const onFlip = (hashDigest: CodemodNodeHashDigest) => {
+	vscode.postMessage({
+		kind: 'webview.global.flipCodemodHashDigest',
+		codemodNodeHashDigest: hashDigest,
+	});
 
+	onFocus(hashDigest);
+};
+
+const TreeView = ({ tree, autocompleteItems, rootPath }: Props) => {
+	/**
+	 * Progress bar
+	 * @TODO hide progress bar logic
+	 */
 	const [executionStack, setExecutionStack] = useState<
-		ReadonlyArray<CodemodHash>
+		ReadonlyArray<CodemodNodeHashDigest>
 	>([]);
-	const [runningRepomodHash, setRunningRepomodHash] =
-		useState<CodemodHash | null>(null);
 
-	useEffect(() => {
-		vscode.postMessage({
-			kind: 'webview.codemods.setState',
-			openedIds: Array.from(state.openedIds),
-			focusedId: state.focusedId,
-		});
-	}, [state]);
+	const [runningRepomodHash, setRunningRepomodHash] =
+		useState<CodemodNodeHashDigest | null>(null);
 
 	const onHalt = useCallback(() => {
 		setRunningRepomodHash(null);
@@ -266,44 +79,30 @@ const TreeView = ({
 		setExecutionStack(stack);
 		vscode.postMessage({
 			kind: 'webview.codemodList.dryRunCodemod',
-			value: hash,
+			// @TODO
+			value: hash as unknown as CodemodHash,
 		});
 	}, [executionStack]);
 
 	const [progress, { progressBar, stopProgress }] = useProgressBar(
 		onHalt,
-		runningRepomodHash,
+		// @TODO make generic
+		runningRepomodHash as CodemodHash | null,
 	);
-
-	useEffect(() => {
-		const handler = (e: MessageEvent<WebviewMessage>) => {
-			const message = e.data;
-
-			if (message.kind === 'webview.codemods.focusCodemod') {
-				dispatch({
-					kind: 'focus',
-					id: message.codemodHashDigest,
-				});
-			}
-		};
-
-		window.addEventListener('message', handler);
-
-		return () => {
-			window.removeEventListener('message', handler);
-		};
-	}, [node]);
 
 	const handleActionButtonClick = useCallback(
 		(action: RunCodemodsCommand) => {
+			// @TODO
+			const actionValue =
+				action.value as unknown as CodemodNodeHashDigest;
 			if (
 				(progress || executionStack.length) &&
 				action.kind === 'webview.codemodList.dryRunCodemod'
 			) {
-				if (executionStack.includes(action.value)) {
+				if (executionStack.includes(actionValue)) {
 					return;
 				}
-				setExecutionStack((prev) => [...prev, action.value]);
+				setExecutionStack((prev) => [...prev, actionValue]);
 				return;
 			}
 
@@ -312,17 +111,25 @@ const TreeView = ({
 		[executionStack, progress],
 	);
 
+	// @TODO move to other component
 	const actionButtons = (
-		node: CodemodTreeNode,
+		node: CodemodNode,
+		actions: RunCodemodsCommand[],
 		doesDisplayShortenedTitle: boolean,
 	) => {
-		return (node.actions ?? []).map((action) => {
+		if (node.kind !== 'CODEMOD') {
+			return [];
+		}
+
+		return (actions ?? []).map((action) => {
+			const actionValue =
+				action.value as unknown as CodemodNodeHashDigest;
 			return (
 				<Popover
 					trigger={
 						<VSCodeButton
 							key={action.kind}
-							className={styles.action}
+							className={s.action}
 							appearance="icon"
 							onClick={(e) => {
 								e.stopPropagation();
@@ -330,15 +137,15 @@ const TreeView = ({
 								if (
 									action.kind ===
 										'webview.codemodList.dryRunCodemod' &&
-									node.modKind === 'repomod'
+									node.codemodKind === 'repomod'
 								) {
-									setRunningRepomodHash(node.id);
+									setRunningRepomodHash(node.hashDigest);
 								}
 							}}
 						>
 							{action.kind ===
 								'webview.codemodList.dryRunCodemod' &&
-								executionStack.includes(action.value) && (
+								executionStack.includes(actionValue) && (
 									<i className="codicon codicon-history mr-2" />
 								)}
 							{doesDisplayShortenedTitle
@@ -348,7 +155,7 @@ const TreeView = ({
 					}
 					popoverText={`${
 						action.kind === 'webview.codemodList.dryRunCodemod' &&
-						executionStack.includes(action.value)
+						executionStack.includes(actionValue)
 							? 'Queued:'
 							: ''
 					} ${action.description}`}
@@ -357,101 +164,59 @@ const TreeView = ({
 		});
 	};
 
-	const renderItem = ({
-		node,
-		depth,
-	}: {
-		node: CodemodTreeNode;
-		depth: number;
-	}) => {
-		const opened = state.openedIds.has(node.id);
+	const getActionButtons = (
+		node: CodemodNode,
+		doesDisplayShortenedTitle: boolean,
+	) => {
+		if (node.kind !== 'CODEMOD') {
+			return [];
+		}
 
-		const icon = getIcon(node.iconName ?? null, opened);
+		if (node.codemodKind === 'repomod' && runningRepomodHash !== null) {
+			return [];
+		}
 
-		const getActionButtons = (doesDisplayShortenedTitle: boolean) => {
-			if (node.modKind === 'repomod' && runningRepomodHash !== null) {
-				return [];
-			}
+		if (
+			// @TODO
+			progress?.codemodHash ===
+				(node.hashDigest as unknown as CodemodHash) &&
+			node.codemodKind === 'executeCodemod'
+		) {
+			return [stopProgress];
+		}
 
-			if (
-				progress?.codemodHash === node.id &&
-				node.modKind === 'executeCodemod'
-			) {
-				return [stopProgress];
-			}
+		const actions = [
+			{
+				title: '✓ Dry Run',
+				shortenedTitle: '✓',
+				description:
+					'Run this codemod without making change to file system',
+				kind: 'webview.codemodList.dryRunCodemod' as const,
+				value: node.hashDigest as unknown as CodemodHash,
+			},
+		];
 
-			return actionButtons(node, doesDisplayShortenedTitle);
-		};
-
-		return (
-			<TreeItem
-				progressBar={
-					progress?.codemodHash === node.id ? progressBar : null
-				}
-				disabled={false}
-				hasChildren={(node.children?.length ?? 0) !== 0}
-				id={node.id}
-				executionPath={node.executionPath}
-				rootPath={rootPath}
-				description={node.description ?? ''}
-				label={node.label ?? ''}
-				icon={icon}
-				depth={depth}
-				kind={node.kind}
-				open={opened}
-				focused={node.id === state.focusedId}
-				autocompleteItems={autocompleteItems}
-				onClick={() => {
-					handleClick(node);
-
-					dispatch({
-						kind: 'flip',
-						id: node.id,
-					});
-				}}
-				onDoubleClick={() => {
-					handleDoubleClick(node);
-				}}
-				actionButtons={(doesDisplayShortenedTitle: boolean) =>
-					getActionButtons(doesDisplayShortenedTitle)
-				}
-				collapse={() => {
-					dispatch({
-						kind: 'close',
-						id: node.id,
-					});
-				}}
-				expand={() => {
-					dispatch({
-						kind: 'open',
-						id: node.id,
-					});
-				}}
-			/>
-		);
+		return actionButtons(node, actions, doesDisplayShortenedTitle);
 	};
 
 	return (
-		<div id="codemodDiscoveryView-treeContainer" tabIndex={0}>
-			<Tree
-				rootPath={rootPath}
-				node={node}
-				renderItem={renderItem}
-				depth={0}
-				openedIds={state.openedIds}
-				hashesForSearch={hashesForSearch}
-				searchingCodemod={userSearchingCodemod}
-				nodeIds={nodeIds}
-				nodesByDepth={nodesByDepth}
-				focusedId={state.focusedId}
-				onFocusNode={(id: CodemodHash) => {
-					dispatch({
-						kind: 'focus',
-						id,
-					});
-				}}
-			/>
-		</div>
+		<IntuitaTreeView<CodemodNodeHashDigest, CodemodNode>
+			{...tree}
+			nodeRenderer={getCodemodNodeRenderer({
+				autocompleteItems,
+				rootPath,
+				onDoubleClick: handleDoubleClick,
+				progressBar: (node: CodemodNode) =>
+					progress?.codemodHash ===
+					(node.hashDigest as unknown as CodemodHash)
+						? progressBar
+						: null,
+				actionButtons: (node, doesDisplayShortenedTitle) =>
+					getActionButtons(node, doesDisplayShortenedTitle),
+			})}
+			onFlip={onFlip}
+			onFocus={onFocus}
+		/>
 	);
 };
